separate file_base model to file
