+++ lib/xeroizer.rb
@@ -25,7 +25,7 @@ require 'xeroizer/record/base_model'
 require 'xeroizer/record/payroll_base_model'
 require 'xeroizer/record/base'
 require 'xeroizer/record/payroll_base'
-require 'xeroizer/record/file_base'
+require 'xeroizer/record/file_base_model'
 require 'xeroizer/configuration'
 
 # Include models
